# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management System API
  description: |
    Backend-приложение на Java (Spring Boot) для управления банковскими картами с поддержкой JWT аутентификации.

    ## Функциональность
    - Создание и управление банковскими картами
    - Просмотр карт с маскированием номеров
    - Переводы между своими картами
    - Аутентификация и авторизация с JWT
    - Ролевая модель доступа (ADMIN, USER)

    ## Безопасность
    - Все эндпоинты, кроме /api/auth/*, требуют JWT токен в заголовке Authorization
    - Номера карт хранятся в зашифрованном виде (AES)
    - При отображении номера маскируются: `**** **** **** 1234`
  version: 1.0.0
  contact:
    email: support@bankcards.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.bankcards.com
    description: Production server
...
paths:
...
  /api/cards/user/{userId}:
    get:
      tags:
        - Cards
      summary: Получение карт пользователя
      description: Возвращает список карт пользователя с пагинацией и опциональной фильтрацией по статусу
      operationId: getUserCards
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED]
          description: Фильтр по статусу карты
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы (начиная с 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Количество элементов на странице
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Некорректный статус карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к картам этого пользователя
...
  /api/transfers/user/{userId}:
    get:
      tags:
        - Transfers
      summary: Получение истории переводов пользователя
      description: Возвращает всю историю переводов для пользователя с пагинацией
      operationId: getUserTransferHistory
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список переводов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTransferResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет доступа к переводам этого пользователя
...
  /api/users:
    get:
      tags:
        - Users
      summary: Получение списка всех пользователей (только для администраторов)
      description: Возвращает список всех пользователей системы с пагинацией
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Количество элементов на странице
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
  /api/users/{userId}:
    get:
      tags:
        - Users
      summary: Получение пользователя по ID (только для администраторов)
      description: Возвращает детальную информацию о пользователе
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
        '404':
          description: Пользователь не найден

    put:
      tags:
        - Users
      summary: Обновление пользователя (только для администраторов)
      description: Обновляет информацию о пользователе
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              email: newemail@example.com
              fullName: New Full Name
              password: newPassword123
      responses:
        '200':
          description: Пользователь успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Некорректные данные (email уже занят)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
        '404':
          description: Пользователь не найден

    delete:
      tags:
        - Users
      summary: Удаление пользователя (только для администраторов)
      description: Удаляет пользователя из системы
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '204':
          description: Пользователь успешно удален
        '401':
          description: Не авторизован
        '403':
          description: Нет прав администратора
        '404':
          description: Пользователь не найден
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при login/register
  schemas:
...
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 1234"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2027-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          description: Статус карты
          example: ACTIVE
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.00
        userId:
          type: integer
          format: int64
          description: ID владельца
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Дата создания
          example: "2025-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления
          example: "2025-01-15T10:30:00"
...
    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID перевода
          example: 1
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          description: Сумма перевода
          example: 100.00
        description:
          type: string
          description: Описание перевода
          example: Transfer between my cards
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
          description: Статус перевода
          example: COMPLETED
        createdAt:
          type: string
          format: date-time
          description: Дата создания
          example: "2025-01-15T10:30:00"
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: johndoe
        email:
          type: string
          format: email
          description: Email
          example: john@example.com
        fullName:
          type: string
          description: Полное имя
          example: John Doe
        roles:
          type: array
          items:
            type: string
          description: Роли пользователя
          example: [ROLE_USER]
        createdAt:
          type: string
          format: date-time
          description: Дата создания
          example: "2025-01-15T10:30:00"
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          description: Новый email (опционально)
          example: newemail@example.com
        fullName:
          type: string
          minLength: 1
          maxLength: 100
          description: Новое полное имя (опционально)
          example: New Full Name
        password:
          type: string
          minLength: 6
          maxLength: 100
          description: Новый пароль (опционально)
          example: newPassword123
    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
          example: 100
        totalPages:
          type: integer
          description: Общее количество страниц
          example: 10
        size:
          type: integer
          description: Размер страницы
          example: 10
        number:
          type: integer
          description: Номер текущей страницы
          example: 0
        first:
          type: boolean
          description: Первая страница?
          example: true
        last:
          type: boolean
          description: Последняя страница?
          example: false
        empty:
          type: boolean
          description: Пустая страница?
          example: false
    PageTransferResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransferResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        page:
          type: integer
          description: Номер страницы
          example: 0
        size:
          type: integer
          description: Размер страницы
          example: 10
        sort:
          type: array
          items:
            type: string
          description: Параметры сортировки
          example: ["createdAt,desc"]
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-01-15T10:30:00"
        status:
          type: integer
          description: HTTP статус код
          example: 400
        error:
          type: string
          description: Описание ошибки
          example: Bad Request
        message:
          type: string
          description: Детальное сообщение об ошибке
          example: Card number already exists
        path:
          type: string
          description: Путь запроса
          example: /api/cards